Shuffle flag for inbox command
Consider a scenario where new papers are added to the inbox at a rapid
pace, and ingesting the inbox is not run to completion. The files at the
end of the asciibetic sort can linger for a long time in the inbox while
newly added papers pass them in the queue. To alleviate the problem,
the flag --shuffle can be specified to process files in a random order.